Change package scope and upgrade libraries
This SDK now used @salling-group/auth and
@salling-group/pagination-traverser intead of the old base SDK. This
means that the SDKs original contract is broken, however, this new
version will deprecate the existing NPM package and be released under the Salling Group organization so it is
not a breaking change.

Furthermore, country filtering has been added to the query builder.